+++ partiql-lang/src/main/kotlin/org/partiql/lang/util/AntlrUtilities.kt
@@ -14,9 +14,18 @@
 
 package org.partiql.lang.util
 
-import com.amazon.ion.IonSystem
+import com.amazon.ion.Decimal
 import com.amazon.ion.IonValue
 import com.amazon.ion.system.IonSystemBuilder
+import com.amazon.ionelement.api.IonElement
+import com.amazon.ionelement.api.ionBool
+import com.amazon.ionelement.api.ionDecimal
+import com.amazon.ionelement.api.ionInt
+import com.amazon.ionelement.api.ionNull
+import com.amazon.ionelement.api.ionString
+import com.amazon.ionelement.api.ionSymbol
+import com.amazon.ionelement.api.loadSingleElement
+import com.amazon.ionelement.api.toIonValue
 import org.antlr.v4.runtime.Token
 import org.antlr.v4.runtime.tree.TerminalNode
 import org.partiql.lang.errors.ErrorCode
@@ -30,15 +39,17 @@ import org.partiql.lang.syntax.TYPE_ALIASES
 import org.partiql.lang.syntax.antlr.PartiQLParser
 import java.math.BigInteger
 
+// workaround until ErrorAndErrorContexts no longer uses IonSystem
+private val ion = IonSystemBuilder.standard().build()
+
 internal fun TerminalNode?.error(
     message: String,
     errorCode: ErrorCode,
     errorContext: PropertyValueMap = PropertyValueMap(),
     cause: Throwable? = null,
-    ion: IonSystem = IonSystemBuilder.standard().build()
 ) = when (this) {
     null -> ParserException(errorCode = errorCode, errorContext = errorContext, cause = cause)
-    else -> this.symbol.error(message, errorCode, errorContext, cause, ion)
+    else -> this.symbol.error(message, errorCode, errorContext, cause)
 }
 
 internal fun Token?.error(
@@ -46,47 +57,50 @@ internal fun Token?.error(
     errorCode: ErrorCode,
     errorContext: PropertyValueMap = PropertyValueMap(),
     cause: Throwable? = null,
-    ion: IonSystem = IonSystemBuilder.standard().build()
 ) = when (this) {
     null -> ParserException(errorCode = errorCode, errorContext = errorContext, cause = cause)
     else -> {
         errorContext[Property.LINE_NUMBER] = this.line.toLong()
         errorContext[Property.COLUMN_NUMBER] = this.charPositionInLine.toLong() + 1
         errorContext[Property.TOKEN_DESCRIPTION] = this.type.getAntlrDisplayString()
-        errorContext[Property.TOKEN_VALUE] = getIonValue(ion, this)
+        errorContext[Property.TOKEN_VALUE] = getIonValue(this)
         ParserException(message, errorCode, errorContext, cause)
     }
 }
 
+internal fun getIonValue(token: Token): IonValue {
+    return getIonElement(token).toIonValue(ion)
+}
+
 /**
- * Returns the corresponding [IonValue] for a particular ANTLR Token
+ * Returns the corresponding [IonElement] for a particular ANTLR Token
  */
-internal fun getIonValue(ion: IonSystem, token: Token): IonValue {
+internal fun getIonElement(token: Token): IonElement {
     val type = token.type
     val text = token.text
     return when {
-        type == PartiQLParser.EOF -> ion.newSymbol("EOF")
-        ALL_OPERATORS.contains(text.toLowerCase()) -> ion.newSymbol(text.toLowerCase())
-        type == PartiQLParser.ION_CLOSURE -> ion.singleValue(text.trimStart('`').trimEnd('`'))
-        type == PartiQLParser.TRUE -> ion.newBool(true)
-        type == PartiQLParser.FALSE -> ion.newBool(false)
-        type == PartiQLParser.NULL -> ion.newNull()
-        type == PartiQLParser.NULLS -> ion.newSymbol("nulls")
-        type == PartiQLParser.MISSING -> ion.newNull()
-        type == PartiQLParser.LITERAL_STRING -> ion.newString(text.trim('\'').replace("''", "'"))
-        type == PartiQLParser.LITERAL_INTEGER -> ion.newInt(BigInteger(text, 10))
+        type == PartiQLParser.EOF -> ionSymbol("EOF")
+        ALL_OPERATORS.contains(text.toLowerCase()) -> ionSymbol(text.toLowerCase())
+        type == PartiQLParser.ION_CLOSURE -> loadSingleElement(text.trimStart('`').trimEnd('`'))
+        type == PartiQLParser.TRUE -> ionBool(true)
+        type == PartiQLParser.FALSE -> ionBool(false)
+        type == PartiQLParser.NULL -> ionNull()
+        type == PartiQLParser.NULLS -> ionSymbol("nulls")
+        type == PartiQLParser.MISSING -> ionNull()
+        type == PartiQLParser.LITERAL_STRING -> ionString(text.trim('\'').replace("''", "'"))
+        type == PartiQLParser.LITERAL_INTEGER -> ionInt(BigInteger(text, 10))
         type == PartiQLParser.LITERAL_DECIMAL -> try {
-            ion.newDecimal(bigDecimalOf(text))
+            ionDecimal(Decimal.valueOf(text))
         } catch (e: NumberFormatException) {
-            throw token.error(e.localizedMessage, ErrorCode.PARSE_EXPECTED_NUMBER, cause = e, ion = ion)
+            throw token.error(e.localizedMessage, ErrorCode.PARSE_EXPECTED_NUMBER, cause = e)
         }
-        type == PartiQLParser.IDENTIFIER_QUOTED -> ion.newSymbol(text.trim('\"').replace("\"\"", "\""))
+        type == PartiQLParser.IDENTIFIER_QUOTED -> ionSymbol(text.trim('\"').replace("\"\"", "\""))
         MULTI_LEXEME_TOKEN_MAP.containsKey(text.toLowerCase().split("\\s+".toRegex())) -> {
             val pair = MULTI_LEXEME_TOKEN_MAP[text.toLowerCase().split("\\s+".toRegex())]!!
-            ion.newSymbol(pair.first)
+            ionSymbol(pair.first)
         }
-        KEYWORDS.contains(text.toLowerCase()) -> ion.newSymbol(TYPE_ALIASES[text.toLowerCase()] ?: text.toLowerCase())
-        else -> ion.newSymbol(text)
+        KEYWORDS.contains(text.toLowerCase()) -> ionSymbol(TYPE_ALIASES[text.toLowerCase()] ?: text.toLowerCase())
+        else -> ionSymbol(text)
     }
 }